Setup script for Void Linux for *GAMING*
Bascially, this will do a lot of the main legwork you need to on your
own for Void Linux.  This is written specifically for a modern AMD PC,
using an AMD CPU and GPU.  It does not take older PC's and Intel CPU's
and Nvidia GPU's into account.

It does install as little proprietary stuff as necessary, but for
gaming, you need proprietary stuff.

I have included in flash here, too, for *my* convenience.  This may
become an outdated version in time, I may update it in time.